--- app.vue
@@ -1,18 +1,0 @@
-<script setup lang="ts">
-useSeoMeta({
-	title: 'Zhaolin Lau',
-	ogTitle: 'Zhaolin Lau',
-	description: 'Web developer with more than 3 years of well-rounded experience with a diploma in the field of Computer Science, extensive knowledge of modern Web techniques and love for Milo. Looking for an opportunity to work and upgrade, as well as being involved in an organization that believes in gaining a competitive edge and giving back to the community.',
-	ogDescription: 'Web developer with more than 3 years of well-rounded experience with a diploma in the field of Computer Science, extensive knowledge of modern Web techniques and love for Milo. Looking for an opportunity to work and upgrade, as well as being involved in an organization that believes in gaining a competitive edge and giving back to the community.',
-	ogImage: 'img/me.jpg',
-	twitterCard: 'summary_large_image',
-})
-</script>
-
-<template>
-	<div>
-		<NuxtLayout>
-			<NuxtPage />
-		</NuxtLayout>
-	</div>
-</template>
