--- layouts/default.vue
@@ -1,5 +1,0 @@
-<template>
-	<NavBar />
-	<slot />
-	<Footer />
-</template>
