--- nuxt.config.ts
@@ -1,5 +1,0 @@
-// https://nuxt.com/docs/api/configuration/nuxt-config
-export default defineNuxtConfig({
-	devtools: { enabled: true },
-	modules: ["@nuxt/image", "nuxt-simple-robots"],
-});
