+++ package.json
@@ -1,24 +1,20 @@
 {
-  "name": "nuxt-app",
+  "name": "zhaolinlau.github.io",
+  "version": "0.0.0",
   "private": true,
   "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "postinstall": "nuxt prepare"
-  },
-  "devDependencies": {
-    "@nuxt/devtools": "latest",
-    "@nuxt/image": "1.0.0-rc.1",
-    "@types/node": "^18.17.1",
-    "nuxt": "^3.6.5",
-    "nuxt-simple-robots": "^3.0.3"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
     "@mdi/font": "^7.2.96",
     "@oruga-ui/oruga-next": "^0.6.0",
     "@oruga-ui/theme-bulma": "^0.2.11",
-    "vue-scrollto": "^2.20.0"
+    "vue": "^3.3.4"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^4.2.3",
+    "vite": "^4.4.4"
   }
 }
