--- plugins/oruga.ts
@@ -1,8 +1,0 @@
-import Oruga from "@oruga-ui/oruga-next";
-import { bulmaConfig } from "@oruga-ui/theme-bulma";
-import "@oruga-ui/theme-bulma/dist/bulma.css";
-import "@mdi/font/css/materialdesignicons.css";
-
-export default defineNuxtPlugin((nuxtApp) => {
-	nuxtApp.vueApp.use(Oruga, bulmaConfig);
-});
