+++ src/assets/huawei-intelligent-world-2030.png
