+++ src/components/about.vue
@@ -1,0 +1,28 @@
+<template>
+	<section class="section" id="about">
+		<div class="container">
+			<p class="title mt-3 has-text-centered">
+				<span class="icon-text">
+					<o-icon size="medium" icon="information-variant-circle-outline"></o-icon>
+					<span>About Me</span>
+				</span>
+			</p>
+			<div class="columns is-multiline is-vcentered">
+				<div class="column is-4">
+					<figure class="image">
+						<img class="is-rounded" width="100" height="100" src="../assets/me.jpg" alt="LAU ZHAO LIN">
+					</figure>
+				</div>
+				<div class="column is-8">
+					<p class="subtitle has-text-centered">Jack of all trades, master of "some"</p>
+					<p class="has-text-justified">
+						Web developer with more than <b>3 years</b> of well-rounded experience with a diploma in the field of
+						<b>Computer Science</b>, extensive knowledge of modern Web techniques and love for <b>Milo</b>. Looking for an
+						opportunity to work and upgrade, as well as being involved in an organization that believes in gaining a
+						competitive edge and giving back to the community.
+					</p>
+				</div>
+			</div>
+		</div>
+	</section>
+</template>
