+++ src/components/hero.vue
@@ -1,0 +1,55 @@
+<template>
+	<section class="hero is-fullheight is-dark is-bold">
+		<div class="hero-body">
+			<div class="container">
+				<p class="subtitle">Hello! I am</p>
+				<p class="title is-1 is-italic">Zhaolin Lau</p>
+				<p class="subtitle is-3">
+					A Passionate Web Developer From Malaysia
+				</p>
+				<div class="buttons">
+					<a class="button is-dark" href="https://github.com/zhaolinlau" target="_blank">
+						<span class="icon-text">
+							<span class="icon">
+								<span class="mdi mdi-github mdi-24px"></span>
+							</span>
+							<span>GitHub</span>
+						</span>
+					</a>
+					<a class="button is-link" href="https://www.facebook.com/zhaolinlau/" target="_blank">
+						<span class="icon-text">
+							<span class="icon">
+								<span class="mdi mdi-facebook mdi-24px"></span>
+							</span>
+							<span>Facebook</span>
+						</span>
+					</a>
+					<a class="button is-danger is-light" href="https://www.instagram.com/zhaolin_lau/" target="_blank">
+						<span class="icon-text">
+							<span class="icon">
+								<span class="mdi mdi-instagram mdi-24px"></span>
+							</span>
+							<span>Instagram</span>
+						</span>
+					</a>
+					<a class="button is-info" href="https://www.linkedin.com/in/zhaolinlau/" target="_blank">
+						<span class="icon-text">
+							<span class="icon">
+								<span class="mdi mdi-linkedin mdi-24px"></span>
+							</span>
+							<span>LinkedIn</span>
+						</span>
+					</a>
+					<a class="button is-info is-inverted" href="https://twitter.com/zhaolin_lau" target="_blank">
+						<span class="icon-text">
+							<span class="icon">
+								<span class="mdi mdi-twitter mdi-24px"></span>
+							</span>
+							<span>Twitter</span>
+						</span>
+					</a>
+				</div>
+			</div>
+		</div>
+	</section>
+</template>
