+++ src/components/navbar.vue
@@ -1,18 +1,12 @@
 <script setup>
 import { ref } from 'vue';
 const showNav = ref(false);
-import VueScrollTo from 'vue-scrollto';
-const scrollToElement = (id) => {
-	VueScrollTo.scrollTo(id, {
-		duration: 500,
-	});
-};
 </script>
 
 <template>
-	<nav class="navbar is-fixed-top has-shadow" role="navigation" aria-label="main navigation" id="nav">
+	<nav class="navbar is-fixed-top has-shadow" role="navigation" aria-label="main navigation">
 		<div class="navbar-brand">
-			<a class="navbar-item" @click="scrollToElement('#nav')">
+			<a class="navbar-item" href="#">
 				<p class="title">Zhaolin Lau</p>
 			</a>
 
@@ -26,35 +20,35 @@ const scrollToElement = (id) => {
 
 		<div class="navbar-menu" :class="{ 'is-active': showNav }">
 			<div class="navbar-end">
-				<a class="navbar-item is-tab" @click="scrollToElement('#about')">
+				<a class="navbar-item is-tab" href="#about">
 					<span class="icon-text">
 						<o-icon icon="information-variant-circle-outline"></o-icon>
 						<span>About</span>
 					</span>
 				</a>
-				<a class="navbar-item is-tab" @click="scrollToElement('#skills')">
+				<a class="navbar-item is-tab" href="#skills">
 					<span class="icon-text">
 						<o-icon icon="code-tags"></o-icon>
 						<span>Skills</span>
 					</span>
 				</a>
-				<a class="navbar-item is-tab" @click="scrollToElement('#portfolio')">
+				<a class="navbar-item is-tab" href="#portfolio">
 					<span class="icon-text">
 						<o-icon icon="folder-file"></o-icon>
 						<span>Portfolio</span>
 					</span>
 				</a>
-				<a class="navbar-item is-tab" @click="scrollToElement('#contact')">
+				<a class="navbar-item is-tab" href="#contact">
 					<span class="icon-text">
 						<o-icon icon="mailbox"></o-icon>
 						<span>Contact</span>
 					</span>
 				</a>
 				<a class="navbar-item" href='https://ko-fi.com/V7V6EWJRF' target='_blank'>
-					<nuxt-img src='img/kofi1.webp' alt='Buy Me a Coffee at ko-fi.com' />
+					<img src='https://storage.ko-fi.com/cdn/kofi1.png?v=3' alt='Buy Me a Coffee at ko-fi.com' />
 				</a>
 				<a class="navbar-item" href="https://www.buymeacoffee.com/zhaolinlau" target="_blank">
-					<nuxt-img src="img/default-yellow.png" alt="Buy Me A Coffee" />
+					<img src="https://cdn.buymeacoffee.com/buttons/v2/default-yellow.png" alt="Buy Me A Coffee">
 				</a>
 			</div>
 		</div>
