+++ src/components/portfolio.vue
@@ -1,0 +1,102 @@
+<template>
+	<section class="section" id="portfolio">
+		<div class="container">
+			<p class="title has-text-centered mt-3">
+				<span class="icon-text">
+					<o-icon size="medium" icon="folder-file"></o-icon>
+					<span>My Portfolio</span>
+				</span>
+			</p>
+
+			<div class="columns is-multiline is-centered">
+				<div class="column is-4">
+					<div class="card">
+						<header class="card-header">
+							<p class="card-header-title">
+								<span>Zhaolin Lau</span>
+							</p>
+						</header>
+						<div class="card-content">
+							<figure class="image">
+								<img src="../assets/zhaolinlau.png" alt="zhaolinlau.png" />
+							</figure>
+						</div>
+						<footer class="card-footer">
+							<a href="https://zhaolinlau.vercel.app/" class="card-footer-item">
+								<span class="icon-text">
+									<o-icon icon="eye"></o-icon>
+									<span>Preview</span>
+								</span>
+							</a>
+							<a href="https://github.com/zhaolinlau/zhaolinlau.github.io" class="card-footer-item" target="_blank">
+								<span class="icon-text">
+									<o-icon icon="github"></o-icon>
+									<span>Source Code</span>
+								</span>
+							</a>
+						</footer>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="card">
+						<header class="card-header">
+							<p class="card-header-title">
+								<span>UTM Foundation & Diploma Consultation</span>
+							</p>
+						</header>
+						<div class="card-content">
+							<figure class="image">
+								<img src="../assets/utmfd.png" alt="utmfd.png" />
+							</figure>
+						</div>
+						<footer class="card-footer">
+							<a href="https://utmfd.vercel.app/" class="card-footer-item" target="_blank">
+								<span class="icon-text">
+									<o-icon icon="eye"></o-icon>
+									<span>Preview</span>
+								</span>
+							</a>
+							<a href="https://github.com/Luetify/UTM-FD" class="card-footer-item" target="_blank">
+								<span class="icon-text">
+									<o-icon icon="github"></o-icon>
+									<span>Source Code</span>
+								</span>
+							</a>
+						</footer>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="card">
+						<header class="card-header">
+							<p class="card-header-title">
+								<span>Huawei Open Day: Intelligent World 2030</span>
+							</p>
+						</header>
+						<div class="card-content">
+							<figure class="image">
+								<img src="../assets/huawei-intelligent-world-2030.png" alt="huawei-intelligent-world-2030.png" />
+							</figure>
+						</div>
+						<footer class="card-footer">
+							<a href="https://huawei-intelligent-world-2030.vercel.app/" class="card-footer-item" target="_blank">
+								<span class="icon-text">
+									<o-icon icon="eye"></o-icon>
+									<span>Preview</span>
+								</span>
+							</a>
+							<a href="https://github.com/zhaolinlau/huawei-intelligent-world-2030" class="card-footer-item"
+								target="_blank">
+								<span class="icon-text">
+									<o-icon icon="github"></o-icon>
+									<span>Source Code</span>
+								</span>
+							</a>
+						</footer>
+					</div>
+				</div>
+			</div>
+		</div>
+	</section>
+</template>
