+++ src/components/skills.vue
@@ -1,0 +1,76 @@
+<template>
+	<section class="section has-text-centered" id="skills">
+		<div class="container">
+			<p class="title mt-3">
+				<span class="icon-text">
+					<o-icon size="medium" icon="code-tags"></o-icon>
+					<span>My Skills</span>
+				</span>
+			</p>
+			<div class="columns is-multiline">
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="language-html5" size="large"></o-icon>
+						<p class="title is-4">HTML</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="language-css3" size="large"></o-icon>
+						<p class="title is-4">CSS</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="language-javascript" size="large"></o-icon>
+						<p class="title is-4">JavaScript</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="language-php" size="large"></o-icon>
+						<p class="title is-4">PHP</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="dolphin" size="large"></o-icon>
+						<p class="title is-4">MySQL</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="bulma" size="large"></o-icon>
+						<p class="title is-4">Bulma</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="bootstrap" size="large"></o-icon>
+						<p class="title is-4">Bootstrap</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="vuejs" size="large"></o-icon>
+						<p class="title is-4">Vue.js</p>
+					</div>
+				</div>
+
+				<div class="column is-4">
+					<div class="box">
+						<o-icon icon="laravel" size="large"></o-icon>
+						<p class="title is-4">Laravel</p>
+					</div>
+				</div>
+			</div>
+		</div>
+	</section>
+</template>
