+++ src/main.js
@@ -1,0 +1,8 @@
+import { createApp } from "vue";
+import App from "./App.vue";
+import Oruga from "@oruga-ui/oruga-next";
+import { bulmaConfig } from "@oruga-ui/theme-bulma";
+import "@oruga-ui/theme-bulma/dist/bulma.min.css";
+import "@mdi/font/css/materialdesignicons.min.css";
+
+createApp(App).use(Oruga, bulmaConfig).mount("#app");
